+++ example/functions/stand-alone.js
@@ -1,5 +1,5 @@
-const Log = require('@perform/lambda-powertools-logger')
-const wrap = require('@perform/lambda-powertools-pattern-basic')
+const Log = require('@dazn/lambda-powertools-logger')
+const wrap = require('@dazn/lambda-powertools-pattern-basic')
 
 module.exports.handler = wrap(async (event, context) => {
   Log.debug('finding the answer to life, the universe and everything...')