+++ packages/lambda-powertools-lambda-client/__tests__/index.js
@@ -6,7 +6,7 @@ AWS.Lambda.prototype.invoke = mockInvoke
 AWS.Lambda.prototype.invokeAsync = mockInvokeAsync
 
 const Lambda = require('../index')
-const CorrelationIds = require('@perform/lambda-powertools-correlation-ids')
+const CorrelationIds = require('@dazn/lambda-powertools-correlation-ids')
 
 global.console.log = jest.fn()