+++ packages/lambda-powertools-pattern-obfuscate/__tests__/index.js
@@ -9,14 +9,14 @@ const MockImplementation = () => {
 const SampleLogging = jest.fn()
 // eslint-disable-next-line no-unused-vars
 const MockSampleLogging = jest.mock(
-  '@perform/lambda-powertools-middleware-sample-logging',
+  '@dazn/lambda-powertools-middleware-sample-logging',
   () => { return SampleLogging.mockImplementation(MockImplementation) }
 )
 
 const CaptureCorrelationIds = jest.fn()
 // eslint-disable-next-line no-unused-vars
 const MockCaptureCorrelationIds = jest.mock(
-  '@perform/lambda-powertools-middleware-correlation-ids',
+  '@dazn/lambda-powertools-middleware-correlation-ids',
   () => { return CaptureCorrelationIds.mockImplementation(MockImplementation) }
 )
 
@@ -27,7 +27,7 @@ afterEach(() => {
   CaptureCorrelationIds.mockClear()
 })
 
-// const LogTimeout = require('@perform/lambda-powertools-middleware-log-timeout')
+// const LogTimeout = require('@dazn/lambda-powertools-middleware-log-timeout')
 const util = require('util')
 
 describe('obfuscate pattern', () => {