+++ packages/lambda-powertools-logger/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.9.0](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.8.3...v1.9.0) (2019-08-23)
+
+
+### Features
+
+* **logger:** make params optional when logging the error stack ([c095376](https://github.com/getndazn/dazn-lambda-powertools/commit/c095376))
+
+
+
+
+
+## [1.8.2](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.8.1...v1.8.2) (2019-07-24)
+
+**Note:** Version bump only for package @dazn/lambda-powertools-logger
+
+
+
+
+
 # [1.8.0](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.7.0...v1.8.0) (2019-07-10)
 
 
@@ -16,7 +35,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 # [1.5.0](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.4.2...v1.5.0) (2019-06-20)
 
-**Note:** Version bump only for package @perform/lambda-powertools-logger
+**Note:** Version bump only for package @dazn/lambda-powertools-logger
 
 
 
@@ -24,7 +43,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 # [1.3.0](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.2.0...v1.3.0) (2019-06-10)
 
-**Note:** Version bump only for package @perform/lambda-powertools-logger
+**Note:** Version bump only for package @dazn/lambda-powertools-logger
 
 
 
@@ -32,7 +51,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [1.1.7](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.1.6...v1.1.7) (2019-05-22)
 
-**Note:** Version bump only for package @perform/lambda-powertools-logger
+**Note:** Version bump only for package @dazn/lambda-powertools-logger
 
 
 
@@ -40,7 +59,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [1.0.1](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.0.0...v1.0.1) (2019-03-04)
 
-**Note:** Version bump only for package @perform/lambda-powertools-logger
+**Note:** Version bump only for package @dazn/lambda-powertools-logger
 
 
 
@@ -65,7 +84,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [0.10.5](https://github.com/getndazn/dazn-lambda-powertools/compare/v0.10.4...v0.10.5) (2019-02-19)
 
-**Note:** Version bump only for package @perform/lambda-powertools-logger
+**Note:** Version bump only for package @dazn/lambda-powertools-logger
 
 
 
@@ -73,7 +92,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [0.10.4](https://github.com/getndazn/dazn-lambda-powertools/compare/v0.10.3...v0.10.4) (2019-02-18)
 
-**Note:** Version bump only for package @perform/lambda-powertools-logger
+**Note:** Version bump only for package @dazn/lambda-powertools-logger
 
 
 
@@ -81,7 +100,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [0.10.3](https://github.com/getndazn/dazn-lambda-powertools/compare/v0.10.2...v0.10.3) (2019-02-18)
 
-**Note:** Version bump only for package @perform/lambda-powertools-logger
+**Note:** Version bump only for package @dazn/lambda-powertools-logger
 
 
 
@@ -89,7 +108,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [0.10.2](https://github.com/getndazn/dazn-lambda-powertools/compare/v0.10.1...v0.10.2) (2019-02-18)
 
-**Note:** Version bump only for package @perform/lambda-powertools-logger
+**Note:** Version bump only for package @dazn/lambda-powertools-logger
 
 
 
@@ -97,4 +116,4 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [0.10.1](https://github.com/getndazn/dazn-lambda-powertools/compare/v0.10.0...v0.10.1) (2019-02-18)
 
-**Note:** Version bump only for package @perform/lambda-powertools-logger
+**Note:** Version bump only for package @dazn/lambda-powertools-logger