+++ packages/lambda-powertools-middleware-correlation-ids/CHANGELOG.md
@@ -3,9 +3,55 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.11.0](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.10.0...v1.11.0) (2019-09-05)
+
+
+### Features
+
+* use API GW request ID as correlation ID ([00e5153](https://github.com/getndazn/dazn-lambda-powertools/commit/00e5153)), closes [#92](https://github.com/getndazn/dazn-lambda-powertools/issues/92)
+
+
+
+
+
+# [1.10.0](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.9.1...v1.10.0) (2019-08-29)
+
+
+### Features
+
+* **firehose:** add middleware support for firehose ([4a8803c](https://github.com/getndazn/dazn-lambda-powertools/commit/4a8803c)), closes [#27](https://github.com/getndazn/dazn-lambda-powertools/issues/27)
+
+
+
+
+
+# [1.9.0](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.8.3...v1.9.0) (2019-08-23)
+
+**Note:** Version bump only for package @dazn/lambda-powertools-middleware-correlation-ids
+
+
+
+
+
+## [1.8.3](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.8.2...v1.8.3) (2019-07-25)
+
+**Note:** Version bump only for package @dazn/lambda-powertools-middleware-correlation-ids
+
+
+
+
+
+## [1.8.2](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.8.1...v1.8.2) (2019-07-24)
+
+**Note:** Version bump only for package @dazn/lambda-powertools-middleware-correlation-ids
+
+
+
+
+
 # [1.8.0](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.7.0...v1.8.0) (2019-07-10)
 
-**Note:** Version bump only for package @perform/lambda-powertools-middleware-correlation-ids
+**Note:** Version bump only for package @dazn/lambda-powertools-middleware-correlation-ids
 
 
 
@@ -13,7 +59,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 # [1.5.0](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.4.2...v1.5.0) (2019-06-20)
 
-**Note:** Version bump only for package @perform/lambda-powertools-middleware-correlation-ids
+**Note:** Version bump only for package @dazn/lambda-powertools-middleware-correlation-ids
 
 
 
@@ -37,7 +83,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 # [1.3.0](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.2.0...v1.3.0) (2019-06-10)
 
-**Note:** Version bump only for package @perform/lambda-powertools-middleware-correlation-ids
+**Note:** Version bump only for package @dazn/lambda-powertools-middleware-correlation-ids
 
 
 
@@ -61,7 +107,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [1.1.7](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.1.6...v1.1.7) (2019-05-22)
 
-**Note:** Version bump only for package @perform/lambda-powertools-middleware-correlation-ids
+**Note:** Version bump only for package @dazn/lambda-powertools-middleware-correlation-ids
 
 
 
@@ -69,7 +115,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [1.1.6](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.1.5...v1.1.6) (2019-05-17)
 
-**Note:** Version bump only for package @perform/lambda-powertools-middleware-correlation-ids
+**Note:** Version bump only for package @dazn/lambda-powertools-middleware-correlation-ids
 
 
 
@@ -99,7 +145,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [1.0.1](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.0.0...v1.0.1) (2019-03-04)
 
-**Note:** Version bump only for package @perform/lambda-powertools-middleware-correlation-ids
+**Note:** Version bump only for package @dazn/lambda-powertools-middleware-correlation-ids
 
 
 
@@ -124,7 +170,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [0.10.5](https://github.com/getndazn/dazn-lambda-powertools/compare/v0.10.4...v0.10.5) (2019-02-19)
 
-**Note:** Version bump only for package @perform/lambda-powertools-middleware-correlation-ids
+**Note:** Version bump only for package @dazn/lambda-powertools-middleware-correlation-ids
 
 
 
@@ -132,7 +178,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [0.10.4](https://github.com/getndazn/dazn-lambda-powertools/compare/v0.10.3...v0.10.4) (2019-02-18)
 
-**Note:** Version bump only for package @perform/lambda-powertools-middleware-correlation-ids
+**Note:** Version bump only for package @dazn/lambda-powertools-middleware-correlation-ids
 
 
 
@@ -140,7 +186,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [0.10.3](https://github.com/getndazn/dazn-lambda-powertools/compare/v0.10.2...v0.10.3) (2019-02-18)
 
-**Note:** Version bump only for package @perform/lambda-powertools-middleware-correlation-ids
+**Note:** Version bump only for package @dazn/lambda-powertools-middleware-correlation-ids
 
 
 
@@ -148,7 +194,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [0.10.2](https://github.com/getndazn/dazn-lambda-powertools/compare/v0.10.1...v0.10.2) (2019-02-18)
 
-**Note:** Version bump only for package @perform/lambda-powertools-middleware-correlation-ids
+**Note:** Version bump only for package @dazn/lambda-powertools-middleware-correlation-ids
 
 
 
@@ -156,4 +202,4 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [0.10.1](https://github.com/getndazn/dazn-lambda-powertools/compare/v0.10.0...v0.10.1) (2019-02-18)
 
-**Note:** Version bump only for package @perform/lambda-powertools-middleware-correlation-ids
+**Note:** Version bump only for package @dazn/lambda-powertools-middleware-correlation-ids