+++ packages/lambda-powertools-middleware-stop-infinite-loop/CHANGELOG.md
@@ -3,9 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.9.0](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.8.3...v1.9.0) (2019-08-23)
+
+**Note:** Version bump only for package @dazn/lambda-powertools-middleware-stop-infinite-loop
+
+
+
+
+
+## [1.8.2](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.8.1...v1.8.2) (2019-07-24)
+
+**Note:** Version bump only for package @dazn/lambda-powertools-middleware-stop-infinite-loop
+
+
+
+
+
 # [1.8.0](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.7.0...v1.8.0) (2019-07-10)
 
-**Note:** Version bump only for package @perform/lambda-powertools-middleware-stop-infinite-loop
+**Note:** Version bump only for package @dazn/lambda-powertools-middleware-stop-infinite-loop
 
 
 
@@ -13,7 +29,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 # [1.5.0](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.4.2...v1.5.0) (2019-06-20)
 
-**Note:** Version bump only for package @perform/lambda-powertools-middleware-stop-infinite-loop
+**Note:** Version bump only for package @dazn/lambda-powertools-middleware-stop-infinite-loop
 
 
 
@@ -21,7 +37,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [1.4.1](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.4.0...v1.4.1) (2019-06-14)
 
-**Note:** Version bump only for package @perform/lambda-powertools-middleware-stop-infinite-loop
+**Note:** Version bump only for package @dazn/lambda-powertools-middleware-stop-infinite-loop