+++ packages/lambda-powertools-pattern-basic/CHANGELOG.md
@@ -3,9 +3,57 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.11.0](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.10.0...v1.11.0) (2019-09-05)
+
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-basic
+
+
+
+
+
+# [1.10.0](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.9.1...v1.10.0) (2019-08-29)
+
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-basic
+
+
+
+
+
+## [1.9.1](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.9.0...v1.9.1) (2019-08-24)
+
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-basic
+
+
+
+
+
+# [1.9.0](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.8.3...v1.9.0) (2019-08-23)
+
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-basic
+
+
+
+
+
+## [1.8.3](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.8.2...v1.8.3) (2019-07-25)
+
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-basic
+
+
+
+
+
+## [1.8.2](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.8.1...v1.8.2) (2019-07-24)
+
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-basic
+
+
+
+
+
 # [1.8.0](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.7.0...v1.8.0) (2019-07-10)
 
-**Note:** Version bump only for package @perform/lambda-powertools-pattern-basic
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-basic
 
 
 
@@ -42,7 +90,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [1.4.2](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.4.1...v1.4.2) (2019-06-20)
 
-**Note:** Version bump only for package @perform/lambda-powertools-pattern-basic
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-basic
 
 
 
@@ -50,7 +98,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 # [1.4.0](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.3.0...v1.4.0) (2019-06-14)
 
-**Note:** Version bump only for package @perform/lambda-powertools-pattern-basic
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-basic
 
 
 
@@ -69,7 +117,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 # [1.2.0](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.1.7...v1.2.0) (2019-05-24)
 
-**Note:** Version bump only for package @perform/lambda-powertools-pattern-basic
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-basic
 
 
 
@@ -77,7 +125,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [1.1.7](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.1.6...v1.1.7) (2019-05-22)
 
-**Note:** Version bump only for package @perform/lambda-powertools-pattern-basic
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-basic
 
 
 
@@ -85,7 +133,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [1.1.6](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.1.5...v1.1.6) (2019-05-17)
 
-**Note:** Version bump only for package @perform/lambda-powertools-pattern-basic
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-basic
 
 
 
@@ -93,7 +141,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [1.1.4](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.1.2...v1.1.4) (2019-04-23)
 
-**Note:** Version bump only for package @perform/lambda-powertools-pattern-basic
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-basic
 
 
 
@@ -101,7 +149,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [1.1.3](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.1.2...v1.1.3) (2019-04-23)
 
-**Note:** Version bump only for package @perform/lambda-powertools-pattern-basic
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-basic
 
 
 
@@ -109,7 +157,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [1.0.1](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.0.0...v1.0.1) (2019-03-04)
 
-**Note:** Version bump only for package @perform/lambda-powertools-pattern-basic
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-basic
 
 
 
@@ -117,7 +165,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 # [1.0.0](https://github.com/getndazn/dazn-lambda-powertools/compare/v0.10.7...v1.0.0) (2019-02-26)
 
-**Note:** Version bump only for package @perform/lambda-powertools-pattern-basic
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-basic
 
 
 
@@ -125,7 +173,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [0.10.5](https://github.com/getndazn/dazn-lambda-powertools/compare/v0.10.4...v0.10.5) (2019-02-19)
 
-**Note:** Version bump only for package @perform/lambda-powertools-pattern-basic
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-basic
 
 
 
@@ -133,7 +181,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [0.10.4](https://github.com/getndazn/dazn-lambda-powertools/compare/v0.10.3...v0.10.4) (2019-02-18)
 
-**Note:** Version bump only for package @perform/lambda-powertools-pattern-basic
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-basic
 
 
 
@@ -141,7 +189,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [0.10.3](https://github.com/getndazn/dazn-lambda-powertools/compare/v0.10.2...v0.10.3) (2019-02-18)
 
-**Note:** Version bump only for package @perform/lambda-powertools-pattern-basic
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-basic
 
 
 
@@ -149,7 +197,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [0.10.2](https://github.com/getndazn/dazn-lambda-powertools/compare/v0.10.1...v0.10.2) (2019-02-18)
 
-**Note:** Version bump only for package @perform/lambda-powertools-pattern-basic
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-basic
 
 
 
@@ -157,4 +205,4 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [0.10.1](https://github.com/getndazn/dazn-lambda-powertools/compare/v0.10.0...v0.10.1) (2019-02-18)
 
-**Note:** Version bump only for package @perform/lambda-powertools-pattern-basic
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-basic