+++ packages/lambda-powertools-pattern-obfuscate/CHANGELOG.md
@@ -3,9 +3,57 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.11.0](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.10.0...v1.11.0) (2019-09-05)
+
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-obfuscate
+
+
+
+
+
+# [1.10.0](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.9.1...v1.10.0) (2019-08-29)
+
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-obfuscate
+
+
+
+
+
+## [1.9.1](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.9.0...v1.9.1) (2019-08-24)
+
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-obfuscate
+
+
+
+
+
+# [1.9.0](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.8.3...v1.9.0) (2019-08-23)
+
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-obfuscate
+
+
+
+
+
+## [1.8.3](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.8.2...v1.8.3) (2019-07-25)
+
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-obfuscate
+
+
+
+
+
+## [1.8.2](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.8.1...v1.8.2) (2019-07-24)
+
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-obfuscate
+
+
+
+
+
 # [1.8.0](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.7.0...v1.8.0) (2019-07-10)
 
-**Note:** Version bump only for package @perform/lambda-powertools-pattern-obfuscate
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-obfuscate
 
 
 
@@ -24,7 +72,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [1.4.2](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.4.1...v1.4.2) (2019-06-20)
 
-**Note:** Version bump only for package @perform/lambda-powertools-pattern-obfuscate
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-obfuscate
 
 
 
@@ -32,7 +80,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 # [1.4.0](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.3.0...v1.4.0) (2019-06-14)
 
-**Note:** Version bump only for package @perform/lambda-powertools-pattern-obfuscate
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-obfuscate
 
 
 
@@ -51,7 +99,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 # [1.2.0](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.1.7...v1.2.0) (2019-05-24)
 
-**Note:** Version bump only for package @perform/lambda-powertools-pattern-obfuscate
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-obfuscate
 
 
 
@@ -59,7 +107,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [1.1.7](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.1.6...v1.1.7) (2019-05-22)
 
-**Note:** Version bump only for package @perform/lambda-powertools-pattern-obfuscate
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-obfuscate
 
 
 
@@ -67,7 +115,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [1.1.6](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.1.5...v1.1.6) (2019-05-17)
 
-**Note:** Version bump only for package @perform/lambda-powertools-pattern-obfuscate
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-obfuscate
 
 
 
@@ -75,7 +123,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [1.1.5](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.1.4...v1.1.5) (2019-04-29)
 
-**Note:** Version bump only for package @perform/lambda-powertools-pattern-obfuscate
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-obfuscate
 
 
 
@@ -83,7 +131,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [1.1.4](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.1.2...v1.1.4) (2019-04-23)
 
-**Note:** Version bump only for package @perform/lambda-powertools-pattern-obfuscate
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-obfuscate
 
 
 
@@ -91,7 +139,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [1.1.3](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.1.2...v1.1.3) (2019-04-23)
 
-**Note:** Version bump only for package @perform/lambda-powertools-pattern-obfuscate
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-obfuscate
 
 
 
@@ -110,7 +158,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [1.1.1](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.1.0...v1.1.1) (2019-03-26)
 
-**Note:** Version bump only for package @perform/lambda-powertools-pattern-obfuscate
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-obfuscate
 
 
 
@@ -129,7 +177,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [1.0.1](https://github.com/getndazn/dazn-lambda-powertools/compare/v1.0.0...v1.0.1) (2019-03-04)
 
-**Note:** Version bump only for package @perform/lambda-powertools-pattern-obfuscate
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-obfuscate
 
 
 
@@ -137,7 +185,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 # [1.0.0](https://github.com/getndazn/dazn-lambda-powertools/compare/v0.10.7...v1.0.0) (2019-02-26)
 
-**Note:** Version bump only for package @perform/lambda-powertools-pattern-obfuscate
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-obfuscate
 
 
 
@@ -145,7 +193,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [0.10.7](https://github.com/getndazn/dazn-lambda-powertools/compare/v0.10.6...v0.10.7) (2019-02-25)
 
-**Note:** Version bump only for package @perform/lambda-powertools-pattern-obfuscate
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-obfuscate
 
 
 
@@ -153,7 +201,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [0.10.6](https://github.com/getndazn/dazn-lambda-powertools/compare/v0.10.5...v0.10.6) (2019-02-19)
 
-**Note:** Version bump only for package @perform/lambda-powertools-pattern-obfuscate
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-obfuscate
 
 
 
@@ -161,7 +209,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [0.10.5](https://github.com/getndazn/dazn-lambda-powertools/compare/v0.10.4...v0.10.5) (2019-02-19)
 
-**Note:** Version bump only for package @perform/lambda-powertools-pattern-obfuscate
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-obfuscate
 
 
 
@@ -169,7 +217,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [0.10.4](https://github.com/getndazn/dazn-lambda-powertools/compare/v0.10.3...v0.10.4) (2019-02-18)
 
-**Note:** Version bump only for package @perform/lambda-powertools-pattern-obfuscate
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-obfuscate
 
 
 
@@ -177,7 +225,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [0.10.3](https://github.com/getndazn/dazn-lambda-powertools/compare/v0.10.2...v0.10.3) (2019-02-18)
 
-**Note:** Version bump only for package @perform/lambda-powertools-pattern-obfuscate
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-obfuscate
 
 
 
@@ -185,7 +233,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [0.10.2](https://github.com/getndazn/dazn-lambda-powertools/compare/v0.10.1...v0.10.2) (2019-02-18)
 
-**Note:** Version bump only for package @perform/lambda-powertools-pattern-obfuscate
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-obfuscate
 
 
 
@@ -193,4 +241,4 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [0.10.1](https://github.com/getndazn/dazn-lambda-powertools/compare/v0.10.0...v0.10.1) (2019-02-18)
 
-**Note:** Version bump only for package @perform/lambda-powertools-pattern-obfuscate
+**Note:** Version bump only for package @dazn/lambda-powertools-pattern-obfuscate